adds quantized implementation of hard sigmoid
Summary:

Adds quantized version of hardsigmoid activation.

Note: not implementing the _ and .out versions is
currently intended, because the implementation changes the scale and
zp and it's nice to not allow the user to specify scale
and zp.  Lmk if we should handle this differently.

Test Plan:

tests
benchmarks

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 8a3fbbf39d7366bd9e968656724b64b406a867c8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34607